Servicebus updates
- Include underscore in package.json
- Replace in-memory schoolCheers with direction checking of socket
rooms. This simplifies potential scaling issues if we need to go
multi-instance for scaling. Also simplifies servicebus/local switching
- socket sessions/rooms now can be optionally stored in service bus
depending on config. This is turned off in Azure now until we decide to
scale across instances.
- made matching changes to client.html to account for backend changes
- added a /listcheers method to dump out current cheer mapping. This is
useful for admin, and to allow a worker to push info into DB for
storage and visualisation.